requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-allauth==64.2.1

django-recaptcha==3.0.0
psycopg2-binary==2.8.5
uwsgi

django-redis==4.12.1
celery==4.4.7
#celery==5.3.1
django-crispy-forms==1.9.0
Pillow==9.5.0
django-imagekit==5.0.0
markdown
django-filter
filetype
django-mptt==0.16.0
django-crispy-forms
requests
django-celery-email
m3u8

# extra nice utilities!
rpdb
tqdm
ipython
flake8
pep8
django-silk
django-debug-toolbar==4.1.0
django-ckeditor==6.6.1
whitenoise
django-cors-headers
